minor refactoring for consistency
