+++ setup.py
@@ -1,15 +1,9 @@
-from distutils.core import setup, find_packages
+from setuptools import setup, find_packages
 
 import os
 import sys
-import pypnm
 
 sys.path.append(os.getcwd())
-
-try:
-    from setuptools import setup
-except ImportError:
-    from distutils.core import setup
 
 setup(
     name='pypnm',
@@ -26,9 +20,12 @@ setup(
         'petsc4py',
         'colorama',
         'python-igraph',
-        'pymetis'
+        'pymetis',
+        'PyTrilinos',
+        'pandas',
+        'mpi4py'
     ],
     author='Karim Khayrat',
     author_email='kkhayrat@gmail.com',
-    description='Pore Network Model'
+    description='Pore Network Modeling package'
 )
